// src/app/models/bar.ts
 // bar.interface.ts
 export interface Bar {
    id_bar?: number;
    nombre_bar: string;
    nombre_ubicacion: string;
    id_ubicacion: number;
    desayuno_horario: string;
    almuerzo_horario: string;
    merienda_horario: string;
    longitud: string;
    latitud: string;
    imagen: string;
   }